Move new game button to header
