+++ src/components/Headers/HomeHeader.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { Text, StyleSheet } from 'react-native';
 
 import MenuButton from '../Buttons/MenuButton';
+import NewGameButton from '../Buttons/NewGameButton';
 import CustomHeader from './CustomHeader';
 
 function HomeHeader({ navigation }) {
@@ -10,6 +11,7 @@ function HomeHeader({ navigation }) {
         <CustomHeader navigation={navigation}
             headerLeft={<MenuButton navigation={navigation} />}
             headerCenter={<Text style={styles.title}>ScorePad</Text>}
+            headerRight={<NewGameButton navigation={navigation} />}
         />
     );
 }
