+++ src/screens/ListScreen.js
@@ -1,62 +1,13 @@
-import React, { memo, useEffect } from 'react';
+import React, { memo } from 'react';
 import { View, StyleSheet } from 'react-native';
-import { Button } from 'react-native-elements';
-import * as Crypto from 'expo-crypto';
-import { useDispatch, useSelector } from 'react-redux';
+import { useSelector } from 'react-redux';
 import Animated, { Layout, Easing } from 'react-native-reanimated';
 
-import { gameSave, selectAllGames } from '../../redux/GamesSlice';
-import { playerAdd } from '../../redux/PlayersSlice';
-import { setCurrentGameId } from '../../redux/SettingsSlice';
+import { selectAllGames } from '../../redux/GamesSlice';
 import GameListItem from '../components/GameListItem';
 
 const ListScreen = ({ navigation }) => {
-    const dispatch = useDispatch();
-
     const gameList = useSelector(state => selectAllGames(state));
-
-    const asyncCreateGame = (dispatch) => new Promise((resolve, reject) => {
-        const player1Id = Crypto.randomUUID();
-        const player2Id = Crypto.randomUUID();
-        const newGameId = Crypto.randomUUID();
-
-        dispatch(playerAdd({
-            id: player1Id,
-            playerName: "Player 1",
-            scores: [0],
-        }));
-        dispatch(playerAdd({
-            id: player2Id,
-            playerName: "Player 2",
-            scores: [0],
-        }));
-        dispatch(gameSave({
-            id: newGameId,
-            title: `Game ${gameList.length + 1}`,
-            dateCreated: Date.now(),
-            roundCurrent: 0,
-            roundTotal: 0,
-            playerIds: [player1Id, player2Id],
-        }));
-
-        dispatch(setCurrentGameId(newGameId));
-        resolve();
-    });
-
-    const addGameHandler = () => {
-        asyncCreateGame(dispatch).then(() => {
-            setTimeout(() => {
-                navigation.navigate("Game");
-            }, 500);
-        });
-    };
-
-    useEffect(() => {
-        if (gameList.length == 0) {
-            addGameHandler();
-
-        }
-    }, [gameList.length]);
 
     const GamesFooter = () => {
         return (
@@ -70,7 +21,6 @@ const ListScreen = ({ navigation }) => {
 
     return (
         <View style={{ flex: 1 }} backgroundColor={'white'}>
-            <Button title="New Game" onPress={addGameHandler} style={styles.newGame} />
             <Animated.FlatList
                 itemLayoutAnimation={Layout.easing(Easing.ease).delay(200)}
                 style={styles.list}
